new project named Infrastructure
This project is filling the gap between Application solution layer and the API solution layer where the application does not know authentication or who is attending
the activity. Hence, Infrastructure will have access to the Application Layer and get the Username of that user to claim that this user is verified, before the user
can join an activity.