Added custom Authorization Policy
IsHostRequirement.cs - responsible for checking whether the user is the host. if the user is the host of an activity, then the user can edit and delete the activity its hosting.

also modified:
IdentityServiceExtensions.cs - added the IsHostRequirement as a requirement to the new policy for handling edit and delete of a hosted activity.
ActivitiesController.cs - added Authorize API endpoints with the corresponding new policy